Support null first and desc
